pulled pojos out of start class
