+++ src/main/java/org/team_trk/domain/BreadOrder.java
@@ -1,5 +1,74 @@
 package org.team_trk.domain;
 
+import java.util.Map;
+
 public class BreadOrder {
+	private String customerId;
+	private String guid;
+	private Date orderDate;
+	private Date deliveryDate;
+	private Map<String, Integer> products;
+
+	public String getCustomerId() {
+		return customerId;
+	}
+
+	public void setCustomerId(String customerId) {
+		this.customerId = customerId;
+	}
+
+	public String getGuid() {
+		return guid;
+	}
+
+	public void setGuid(String guid) {
+		this.guid = guid;
+	}
+
+	public Date getOrderDate() {
+		return orderDate;
+	}
+
+	public void setOrderDate(Date orderDate) {
+		this.orderDate = orderDate;
+	}
+
+	public Date getDeliveryDate() {
+		return deliveryDate;
+	}
+
+	public void setDeliveryDate(Date deliveryDate) {
+		this.deliveryDate = deliveryDate;
+	}
+
+	public Map<String, Integer> getProducts() {
+		return products;
+	}
+
+	public void setProducts(Map<String, Integer> products) {
+		this.products = products;
+	}
+
+	class Date {
+		private int day;
+		private int hour;
+
+		public int getDay() {
+			return day;
+		}
+
+		public void setDay(int day) {
+			this.day = day;
+		}
+
+		public int getHour() {
+			return hour;
+		}
+
+		public void setHour(int hour) {
+			this.hour = hour;
+		}
+
+	}
 
 }
