+++ src/main/java/org/team_trk/domain/Packaging.java
@@ -1,0 +1,22 @@
+package org.team_trk.domain;
+
+public class Packaging {
+	private int boxingTemp;
+	private int breadsPerBox;
+
+	public int getBoxingTemp() {
+		return boxingTemp;
+	}
+
+	public void setBoxingTemp(int boxingTemp) {
+		this.boxingTemp = boxingTemp;
+	}
+
+	public int getBreadsPerBox() {
+		return breadsPerBox;
+	}
+
+	public void setBreadsPerBox(int breadsPerBox) {
+		this.breadsPerBox = breadsPerBox;
+	}
+}
